requirements.txt
anyio==4.1.0
certifi==2023.11.17
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
python-telegram-bot==20.6
sniffio==1.3.0
